Add search of bad customers 	modified:   src/main/java/com/github/thecoolersuptelov/aikamsoftjavajunior/DAO/CustomerRepository.java 	modified:   src/main/java/com/github/thecoolersuptelov/aikamsoftjavajunior/DTO/Input/Criteria.java 	modified:   src/main/java/com/github/thecoolersuptelov/aikamsoftjavajunior/Services/Search.java

+++ src/main/java/com/github/thecoolersuptelov/aikamsoftjavajunior/DTO/Input/Criteria.java
@@ -23,9 +23,9 @@ public class Criteria implements Serializable {
     @JsonProperty("minTimes")
     public long minTimes;
     @JsonProperty("minExpenses")
-    public int minExpenses;
+    public float minExpenses;
     @JsonProperty("maxExpenses")
-    public int maxExpenses;
+    public float maxExpenses;
     @JsonProperty("badCustomers")
     public int badCustomers;
 }
